Add job feature is working

+++ routes/resume.js
@@ -57,16 +57,43 @@ router.post('/:id/status', checkPermission(), async (req, res) => {
 
 router.post('/:id/delete', checkPermission(), async (req, res) => {
   const resume = await ResumeDAL.getResumeByID(req.params.id)
+  if (!resume) return res.status(404).render('core/not_found', { layout: 'core_layout' })
+  if (resume.user.toString() !== req.user._id.toString()) return res.status(404).render('core/not_found', { layout: 'core_layout' })
+
   for (const education of resume.educations) await EducationDAL.deleteEducation(education._id)
   for (const job of resume.jobs) await JobDAL.deleteJob(job._id)
   await ResumeDAL.deleteResume(resume._id)
   res.redirect('/resume')
 })
 
+router.post('/:id/add_job', checkPermission(), async (req, res) => {
+  const resume = await ResumeDAL.getResumeByID(req.params.id)
+  if (!resume) return res.status(404).render('core/not_found', { layout: 'core_layout' })
+  if (resume.user.toString() !== req.user._id.toString()) return res.status(404).render('core/not_found', { layout: 'core_layout' })
+
+  const data = { ...req.body }
+
+  if (data.title === '') delete data.title
+  if (data.company === '') delete data.company
+  if (data.location === '') delete data.location
+  if (data.duties === '') delete data.duties
+  if (data.achievements === '') data.achievements = undefined
+  data.from_date = Date.parse(req.body.from_date) || undefined
+  if (data.from_date === undefined) delete data.from_date
+  data.to_date = Date.parse(req.body.to_date) || undefined
+  if (data.to_date === undefined) delete data.to_date
+
+  const job = await JobDAL.addJob(data)
+  await ResumeDAL.addJobToResume(resume._id, job._id)
+
+  res.redirect(`/resume/${req.params.id}`)
+})
+
 router.get('/:id', checkPermission(), async (req, res) => {
   const resume = await ResumeDAL.getResumeByID(req.params.id)
   if (!resume) return res.status(404).render('core/not_found', { layout: 'core_layout' })
   if (resume.user.toString() !== req.user._id.toString()) return res.status(404).render('core/not_found', { layout: 'core_layout' })
+  console.log(resume)
   res.render('resume/show', { resume })
 })
